client/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
protobuf==4.24.4
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
